fix: bugfixes, input validation, resource leak fixes, and DB transaction safety
- YoutubeManager: use try-with-resources and disconnect HTTP connections
- CustomRoomLayout: set needsUpdate=false only after successful DB write
- RoomChatMessageBubbles: fix duplicate IDs (46-53 all had ID 45), use ConcurrentHashMap
- RoomLayout: handle invalid chars in heightmap parsing
- RoomTile: initialize units to empty set instead of null
- CatalogBuyItemEvent: clamp purchase count to [1, 100]
- FloorPlanEditorSaveEvent: fix permission check, boundary validation, null tile check, strip \r
- RequestGuildBuyEvent: byte to int to prevent overflow
- RequestDeleteRoomEvent: wrap deletion in DB transaction with rollback
- SearchRoomsEvent: use substring() instead of split(), add length checks
- RoomSettingsSaveEvent: clamp all room settings to valid ranges
- TradeOfferMultipleItemsEvent: limit trade count to 100
- PluginManager: close URLClassLoader and stream in finally block

+++ Emulator/src/main/java/com/eu/habbo/messages/incoming/guilds/RequestGuildBuyEvent.java
@@ -72,7 +72,7 @@ public class RequestGuildBuyEvent extends MessageHandler {
 
                     StringBuilder badge = new StringBuilder();
 
-                    byte base = 1;
+                    int base = 1;
 
                     while (base < count) {
                         int id = this.packet.readInt();
